Move StoreDefaults to BaseStore class
Previously, `StoreDefaults` was a simple object that was mixed into each
individual store. This caused the issue that the `collection` property,
an array the holds records for each store, was shared between all
stores.

My proposed fix is to create a `BaseStore` class and have each store be
a singleton instance of a class that inherits from it.

I don't think this caused any regressions, but we'll have to smoke test
to make sure.